Save System fixed
Script PlayerData was deleted due to the scrip was merged to SaveSystem script.
Methods writted in multiple lines, also I deleted the variables that we are not using like the player position and the score of the player.
