+++ GOLF/Assets/_Project/Scripts/Save System/ISaveSource.cs
@@ -6,10 +6,6 @@ namespace Golf
     {
         void Save(GameData data);
         GameData Load();
-        int GetScore();
-        void SetScore(int score);
-        Vector2 GetPlayerPosition();
-        void SetPlayerPosition(Vector2 position);
         float GetVolume();
         void SetVolume(float volume);
     }
