--- GOLF/Assets/_Project/Scripts/Save System/PlayerData.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: dac3b03e57cf5ad409ae540ab4acd22e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
